+++ src/main/java/cmpt276/pg6/realtorest/models/AdminRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface AdminRepository extends JpaRepository<User, Integer>{
+public interface AdminRepository extends JpaRepository<Admin, Integer>{
     List<Admin> findByEmailAndPassword(String email, String password);
 }
