+++ src/main/resources/templates/protected.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>Home Page</title>
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous" />
+        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous" defer></script>
+        <link href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@6.4.2/css/all.css" rel="stylesheet" />
+        <!-- Playfair Display Font -->
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+        <link href="https://fonts.googleapis.com/css2?family=Playfair+Display:ital,wght@0,400..900;1,400..900&display=swap" rel="stylesheet" />
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-icons/1.4.0/font/bootstrap-icons.min.css" />
+        <link rel="stylesheet" href="/css/homepage.css" />
+    </head>
+    <body>
+  
+        <!-- NAVIGATION BAR -->
+        <nav class="navbar navbar-expand-lg bg-body-tertiary">
+            <section class="realtorest-title">
+                <a th:href="@{/}">
+                    <p class="playfair-semibold" style="color: white">Realtorest</p>
+                </a>
+            </section>
+            <section class="navbar-categories ms-auto">
+                <!-- Check to see if the admin signed in or not -->
+
+                <th:block th:if="${admin != null}">
+                    <a th:href="@{/viewUsers}" class="navbar-category"><p>View All Users</p></a>
+                    <a th:href="@{/edit-listings}" class="navbar-category"><p>Edit Listings</p></a>
+
+                    <a th:href="@{/logout}" class="navbar-category"><p>Admin Logout</p></a>
+                    
+                </th:block>
+            </section>
+        </nav>
+
+        <!-- INTRODUCTION -->
+        <section class="banner">
+            <section class="container">
+                <p class="title playfair-semibold">Welcome to Realtorest Admin Page</p>
+                <a th:href="@{/edit-listing}"><button type="button" class="btn btn-primary view-listing">Edit Listings</button></a>
+            </section>
+        </section>
+    </body>
+</html>
