+++ src/main/resources/templates/users/registerAdmin.html
@@ -1,0 +1,91 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <title>Registration</title>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width" />
+        <link rel="stylesheet" type="text/css" href="/css/register.css" />
+        <!-- Bootstrap CSS -->
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous" />
+        <!-- FontAwesome -->
+        <link href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@6.4.2/css/all.css" rel="stylesheet" />
+        <!-- Google Fonts & Custom CSS -->
+        <link href="https://fonts.googleapis.com/css2?family=Pacifico&family=Dancing+Script&family=Lobster&family=Kaushan+Script&family=Great+Vibes&family=Amatic+SC&family=Shadows+Into+Light&family=Satisfy&family=Courgette&family=Tangerine&display=swap" rel="stylesheet" />
+        <link href="https://fonts.googleapis.com/css?family=Raleway:100" rel="stylesheet" type="text/css" />
+        <link href="https://fonts.googleapis.com/css?family=Yanone+Kaffeesatz:200" rel="stylesheet" type="text/css" />
+        <link href="https://fonts.googleapis.com/css2?family=Playfair+Display:ital,wght@0,400..900;1,400..900&display=swap" rel="stylesheet" />
+        <link rel="stylesheet" href="/css/homepage.css" />
+        <link rel="stylesheet" href="/css/login.css" />
+    </head>
+
+    <body>
+        <section class="registration-section py-5">
+            <div class="container">
+                <div class="col-mb-5 register-container">
+                    <div class="register-form mb-3">
+                        <h2 class="text-center">ADMIN REGISTRATION</h2>
+                        <form action="/users/addAdmin" method="post" onsubmit="return validateForm()">
+                            <input type="hidden" name="redirectUrl" value="/" />
+                            <div class="input-wrapper">
+                                <input type="text" class="username-input" placeholder="" name="username" required />
+                                <label class="input-label">Username</label>
+                            </div>
+                            <div class="input-wrapper">
+                                <input type="email" class="email-input" placeholder="" name="email" required />
+                                <label class="input-label">Email</label>
+                            </div>
+                            <div class="input-wrapper">
+                                <input type="password" class="password-input" id="password" placeholder="" name="password" required />
+                                <label class="input-label">Password</label>
+                            </div>
+                            <div class="input-wrapper">
+                                <input type="password" class="confirm-password-input" id="confirm_password" placeholder="" name="confirm_password" required />
+                                <label class="input-label">Confirm Password</label>
+                            </div>
+                            <button type="submit">Register</button>
+                        </form>
+                    </div>
+                </div>
+            </div>
+        </section>
+
+        <!-- NAVIGATION BAR -->
+        <nav class="navbar navbar-expand-lg bg-body-tertiary">
+            <section class="realtorest-title">
+                <a th:href="@{/}">
+                    <p class="playfair-semibold" style="color: white">Realtorest</p>
+                </a>
+            </section>
+            <section class="navbar-categories ms-auto">
+                <a class="navbar-category" data-bs-toggle="offcanvas" data-bs-target="#offcanvasTop" aria-controls="offcanvasTop">
+                    <p>About Us</p>
+                </a>
+                <a href="" class="navbar-category">
+                    <p>Buy</p>
+                </a>
+                <a href="" class="navbar-category">
+                    <p>Rent</p>
+                </a>
+                <a href="" class="navbar-category">
+                    <p>Sell</p>
+                </a>
+                <a href="" class="navbar-category">
+                    <p>Sign In</p>
+                </a>
+            </section>
+        </nav>
+
+        <script>
+            function validateForm() {
+                var password = document.getElementById("password").value
+                var confirm_password = document.getElementById("confirm_password").value
+
+                if (password !== confirm_password) {
+                    alert("Passwords do not match!")
+                    return false
+                }
+                return true
+            }
+        </script>
+    </body>
+</html>
